Update version to 0.14.3 (#2063)
Welcome to pgrx v0.14.3.  

This point release fixes some issues discovered with the new `cargo pgrx
regress` command. Additionally, bindings have been regenerated against
the latest Postgres point releases that dropped on May 8th, 2025, though
no new headers have been included.


As always, please update with `cargo install cargo-pgrx --version 0.14.3
--locked` and update your extension `Cargo.toml` files with `cargo pgrx
upgrade`.

## What's Changed

* Fixes to `cargo pgrx regress` by @eeeebbbbrrrr in
https://github.com/pgcentralfoundation/pgrx/pull/2062
   1) Pressing `<ENTER>` to "Accept[Y, n]?" a test no longer panics
2) While we still `git add <new expected/test_name.out>` files, we no
longer do it to the expected files that we promote in the face of a test
failure
3) Fix a bug where we'd think there's a `test_name.out` file to copy to
`expected/` when that's not actually true
4) The `setup.sql` test is now treated as a normal test, and the only
special handling around is that we'll only run it if we detect we need
to
   5) Cleanup test run output to be consistent between Postgres versions
6) Set `PGRX_REGRESS_TESTING=1` so an extension running under the
regression test suite can detect it


**Full Changelog**:
https://github.com/pgcentralfoundation/pgrx/compare/v0.14.2...v0.14.3