Example : register PostgreSQL hooks (#2125)
Introducing a basic safety catch example to demonstrate how to implement
and register PostgreSQL hooks.

* A post_parse_analyze hook will prevent users from using the DELETE
command without a WHERE clause.
* A process_utility hook will prevent non-superuser from using TRUNCATE.

Following a discussion on the discord channel, I wrote this code to
understand how to migrate from PGRX 0.14 to PGRX 0.16 and replace the
former PgHooks trait that was removed recently.

I thought this could help other extension developers in the same
situation...

Let me know if that's the proprer way to do this....

Co-authored-by: damien clochard <damien.clochard@dalibo.com>