Replace `impl AsRef<T>` with `&T` (#2182)
A pattern that was popular some years ago, when `impl Trait` arguments
were first introduced, was to make everything an `impl Trait` arg. This
can be undesirable because it weakens type inference and spreads
whatever indirection is used throughout the codebase, demanding repeat
invocations of `.as_ref()` or whatever. Unfortunately, pgrx was first
made in the period of that being en vogue.

The alternative, of requiring things like `Option<&T>`, does demand
using `Option::as_deref` a bit, but fortunately that convenience exists
nowadays. It also means that the indirection is handled once at the top
level and not addressed after that, which is more appropriate for
internals. ~~None~~ **Almost** none of the relevant functions in pgrx
are considered public API, after all, as they are found throughout
cargo-pgrx and pgrx-sql-entity-graph.

Remove almost all instances of `impl AsRef`, usually with `Path` or
`str`, and replace them with `&Path` or `&str` as appropriate. This
deletes a few lines from the codebase outright, and in those that
remain, it usually simplifies the existing code. In a couple cases, the
benefit involves actually removing `clone`s of data!

The remaining instances linger with a note about their shame.