Manually bind `static MyProcPort` and `struct Port` (#2162)
This is my proposed alternative to pgcentralfoundation/pgrx#2117
building on @usamoi's suggestion. In order to enable interacting with
`static MyProcPort`, we introduce a definition of the `struct Port` type
in `mod libpq` that is available on all versions of Postgres. Several
fields are left private to discourage access to them, as in general
enabling access to Postgres's cryptography APIs for authentication and
authorization is outside the desired project scope of PGRX.

In addition, in one (but only one) version, Postgres 17, the result of
`size_of::<Port>()` is not even guaranteed to be correct. This is
because we don't immediately know, without implementing
`cfg(accessible(..))`, whether or not these fields should exist or not.
Because of these difficulties, we leave those dubious fields deprecated
on Postgres 17 so that no one interacts with them until such time as we
gain more confidence in that.

Fixes pgcentralfoundation/pgrx#2157

---------

Co-authored-by: Chris Bandy <bandy.chris@gmail.com>
Co-authored-by: usamoi <usamoi@outlook.com>