Use diagnostic attrs for ABI/SQL-related traits (#2203)
We use the attributes to hide `FunctionMetadata` from diagnostic output,
as it is a "behind the scenes" trait, and emphasize `SqlTranslatable`.
We also comment briefly on why `ArgAbi<'_>` and `RetAbi` are like they
are. This significantly improves the diagnostics for all of these by
massively cutting down on noise generated while flailing around.