Add `MemCx<'current, T>` arguments and `PBox<'current, T>` returns (#2210)
tl;dr: `caller_picks_lifetime + macro_expansion == unsoundness`,
so prevent that by adding API for new, lifetime-bound allocations.

For functions exposed to PostgreSQL via `pg_extern`, signatures often
resemble these:

```rust
#[pg_extern]
fn do_something(
    floats: Array<'a, f32>,
    hms: Time,
    day: TimestampTz
) -> Vec<TimestampTz> {
```

```sql
-- pgrx will generate something like this
CREATE FUNCTION do_something(
    floats real[],
    hms time,
    day timestamptz
) RETURNS timestamptz[]
```


In SQL, we accept an array and return an array of a different type, but
in Rust, we return a `Vec`. This is because we lack API to allocate and
return a fresh `Array<'a, T>`s. Behind the scenes, we *do* allocate an
`ArrayType` at the translation boundary, unseen to a Rust programmer.
That, plus the Vec, means two allocations just to mutate an array, even
for a fixed-size type!

Unfortunately, pgrx-offered type translations like `Array<'a, T>`,
cannot have a lifetime *and* expose a new *safe* function to make it
from lifetime-free values! Such API must have the `unsafe` requirement
of access-exclusivity for its lifetime, described in functions like
`slice::from_raw_parts_mut`[^1], because the lifetime becomes
caller-selected. This makes it easy to use a `'static` lifetime with
`pg_extern` and expand `unsafe` code that trusts that lifetime.

A `'static` lifetime is always wrong for `palloc`ated types, even with
`TopMemoryContext`, since it dies whenever Postgres resets the context
and Postgres does have conditions under which it performs such a global
reset. The result is various extremely subtle soundness problems are
possible, yet the root cause lies in code that most never read. We can
solve this by denying the possibility of a caller-selected lifetime.
This will require removing some types, ultimately, but first we need an
API to migrate to.

This new API is `MemCx<'mcx, T>` and `PBox<'mcx, T>`. Importantly, there
is no *safe* route to creating `MemCx<'mcx, T>` with a caller-selected
lifetime. One can only reach `MemCx<'mcx, T>` safely as an argument:
either pass a closure to `memcx::current_context` or call a new
`pg_extern` function via SQL. Either way gives you `&MemCx<'current>`.
These let you create a lifetime-bound allocation that can live long
enough to return it.

We can already return fixed-size pass-by-reference types, so right now
`PBox` is only an optimization. It allows not putting a type on the
stack and then immediately moving it directly into a heap allocation.
More importantly, it enables writing tests, so it is important ground
work for returning de facto unsized values, as they can hide behind
`PBox` which is always `Sized`.

[^1]: https://doc.rust-lang.org/std/slice/fn.from_raw_parts_mut.html