Allow explicit handling of allocation errors (#2226)
Right now, we have several fallible constructors for Postgres types.
When this is an allocation error, we don't allow handling it.
As I expect there to be more such fallible constructors in the future,
and transient memory contexts can be arbitrarily limited in size,
I think we should expose this problem in otherwise-fallible code.
We can retain some infallible constructors elsewhere, though.
If nothing else, this can allow better diagnostics via `track_caller`.

Resolves #2225